reformatted the project. Updated Maven, Added requirements that is needed by core.
