+++ api/src/main/java/org/kakara/kvanilla/api/KConstants.java
