+++ api/src/main/java/org/kakara/kvanilla/api/bases/StoneBase.java
