+++ mod/src/main/java/org/kakara/kvanilla/mod/KVanilla.java
@@ -2,6 +2,8 @@ package org.kakara.kvanilla.mod;
 
 import org.kakara.core.game.Item;
 import org.kakara.core.game.ItemManager;
+import org.kakara.core.mod.ModType;
+import org.kakara.core.mod.annotations.ModInfo;
 import org.kakara.core.mod.annotations.OnDisable;
 import org.kakara.core.mod.annotations.OnEnable;
 import org.kakara.core.mod.annotations.Reload;
@@ -12,7 +14,9 @@ import org.kakara.kvanilla.mod.items.dirt.CoarseDirt;
 import org.kakara.kvanilla.mod.items.dirt.Dirt;
 import org.kakara.kvanilla.mod.items.dirt.FineDirt;
 import org.kakara.kvanilla.mod.items.dirt.GrassyDirt;
+import org.slf4j.Logger;
 
+@ModInfo(name = "KVanilla", description = "The Vanilla Game for Kakara", authors = "The Kakara Development Team", modType = ModType.REGULAR, version = "1.0-SNAPSHOT")
 public class KVanilla extends GameMod {
     private static KVanilla instance;
 
@@ -55,5 +59,11 @@ public class KVanilla extends GameMod {
 
     public static Item get(String name) {
         return getInstance().getKakaraCore().getItemManager().getItem(name);
+    }
+
+    @Override
+    public Logger getLogger() {
+        //Will be deleted later. Just let it sit here to make intellij happy
+        return null;
     }
 }
