+++ mod/src/main/resources/mod.properties
@@ -1,0 +1,1 @@
+main.class=org.kakara.kvanilla.mod.KVanilla
