+++ mod/src/main/resources/resources/textures/16/coarse_dirt.png
