+++ mod/src/main/resources/resources/textures/16/stone.png
